refactor: unify session context imports and remove unused components for cleaner code
