+++ app/[locale]/(admin)/admin/setting/roles/change-role-authority-dialog.tsx
@@ -9,7 +9,7 @@ import { toast } from '@/components/ui/sonner';
 import { ToggleGroup, ToggleGroupItem } from '@/components/ui/toggle-group';
 
 import { revalidate } from '@/action/action';
-import { useSession } from '@/context/session-context.client';
+import { useSession } from '@/context/session-context';
 import useClientApi from '@/hooks/use-client';
 import useQueriesData from '@/hooks/use-queries-data';
 import { groupBy } from '@/lib/utils';
