+++ app/[locale]/(user)/upload/schematic/page.tsx
@@ -20,7 +20,7 @@ import { Textarea } from '@/components/ui/textarea';
 import UserCard from '@/components/user/user-card';
 
 import { IMAGE_PREFIX } from '@/constant/constant';
-import { useSession } from '@/context/session-context.client';
+import { useSession } from '@/context/session-context';
 import useClientApi from '@/hooks/use-client';
 import { useI18n } from '@/i18n/client';
 import { createSchematic, getSchematicPreview } from '@/query/schematic';
@@ -31,6 +31,8 @@ import { CreateSchematicRequest, CreateSchematicSchema } from '@/types/schema/zo
 
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation } from '@tanstack/react-query';
+
+/* eslint-disable @next/next/no-img-element */
 
 /* eslint-disable @next/next/no-img-element */
